Frequency of Value Combinations in Data Frame Columns

Frequency of Value Combinations in Data Frame Columns

fieldSummary(x, groupBy = names(x)[-1L], lengthColumn = "", na = "Unknown")

Arguments

x

data frame

groupBy

vector of character naming the columns (fields) in x to be included in the evaluation. Default: names of all columns in x except the first one (assuming it could be an ID column).

lengthColumn

optional. Name of column in x to be summed up

na

optional. Value to be treated as NA. Default: "Unknown"

Examples

n <- 1000L
sample_replace <- function(x, ...) sample(x, size = n, replace = TRUE, ...)
x <- data.frame(
  pipe_id = 1:n,
  material = sample_replace(c("clay", "concrete", "other")),
  age_cat = sample_replace(c("young", "old")),
  length = as.integer(rnorm(n, 50)),
  stringsAsFactors = FALSE
)

fieldSummary(x)
#>    material age_cat length Count Percentage
#> 1      clay     old     47     2        0.2
#> 2  concrete     old     47     1        0.1
#> 3     other     old     47     1        0.1
#> 4      clay   young     47     1        0.1
#> 5  concrete   young     47     1        0.1
#> 6     other   young     47     6        0.6
#> 7      clay     old     48    21        2.1
#> 8  concrete     old     48    29        2.9
#> 9     other     old     48    22        2.2
#> 10     clay   young     48    25        2.5
#> 11 concrete   young     48    26        2.6
#> 12    other   young     48    25        2.5
#> 13     clay     old     49    52        5.2
#> 14 concrete     old     49    64        6.4
#> 15    other     old     49    67        6.7
#> 16     clay   young     49    43        4.3
#> 17 concrete   young     49    56        5.6
#> 18    other   young     49    62        6.2
#> 19     clay     old     50    56        5.6
#> 20 concrete     old     50    58        5.8
#> 21    other     old     50    48        4.8
#> 22     clay   young     50    60        6.0
#> 23 concrete   young     50    53        5.3
#> 24    other   young     50    61        6.1
#> 25     clay     old     51    22        2.2
#> 26 concrete     old     51    24        2.4
#> 27    other     old     51    27        2.7
#> 28     clay   young     51    21        2.1
#> 29 concrete   young     51    22        2.2
#> 30    other   young     51    22        2.2
#> 31     clay     old     52     4        0.4
#> 32 concrete     old     52     6        0.6
#> 33    other     old     52     2        0.2
#> 34     clay   young     52     4        0.4
#> 35 concrete   young     52     3        0.3
#> 36    other   young     52     2        0.2
#> 37 concrete     old     53     1        0.1
fieldSummary(x, "age_cat")
#>   age_cat Count Percentage
#> 1     old   507       50.7
#> 2   young   493       49.3
fieldSummary(x, "material")
#>   material Count Percentage
#> 1     clay   311       31.1
#> 2 concrete   344       34.4
#> 3    other   345       34.5
fieldSummary(x, "material", lengthColumn = "length")
#>   material Count length Percentage
#> 1     clay   311  15413       31.1
#> 2 concrete   344  17031       34.4
#> 3    other   345  17063       34.5