Frequency of Value Combinations in Data Frame Columns

Frequency of Value Combinations in Data Frame Columns

Usage

fieldSummary(x, groupBy = names(x)[-1L], lengthColumn = "", na = "Unknown")

Arguments

x

data frame

groupBy

vector of character naming the columns (fields) in x to be included in the evaluation. Default: names of all columns in x except the first one (assuming it could be an ID column).

lengthColumn

optional. Name of column in x to be summed up

na

optional. Value to be treated as NA. Default: "Unknown"

Examples

n <- 1000L
sample_replace <- function(x, ...) sample(x, size = n, replace = TRUE, ...)
x <- data.frame(
  pipe_id = 1:n,
  material = sample_replace(c("clay", "concrete", "other")),
  age_cat = sample_replace(c("young", "old")),
  length = as.integer(rnorm(n, 50)),
  stringsAsFactors = FALSE
)

fieldSummary(x)
#>    material age_cat length Count Percentage
#> 1      clay     old     47     2        0.2
#> 2  concrete     old     47     1        0.1
#> 3     other     old     47     8        0.8
#> 4      clay   young     47     2        0.2
#> 5  concrete   young     47     1        0.1
#> 6     other   young     47     1        0.1
#> 7      clay     old     48    19        1.9
#> 8  concrete     old     48    29        2.9
#> 9     other     old     48    12        1.2
#> 10     clay   young     48    23        2.3
#> 11 concrete   young     48    15        1.5
#> 12    other   young     48    25        2.5
#> 13     clay     old     49    70        7.0
#> 14 concrete     old     49    61        6.1
#> 15    other     old     49    57        5.7
#> 16     clay   young     49    70        7.0
#> 17 concrete   young     49    54        5.4
#> 18    other   young     49    58        5.8
#> 19     clay     old     50    52        5.2
#> 20 concrete     old     50    55        5.5
#> 21    other     old     50    58        5.8
#> 22     clay   young     50    54        5.4
#> 23 concrete   young     50    59        5.9
#> 24    other   young     50    52        5.2
#> 25     clay     old     51    24        2.4
#> 26 concrete     old     51    25        2.5
#> 27    other     old     51    19        1.9
#> 28     clay   young     51    29        2.9
#> 29 concrete   young     51    24        2.4
#> 30    other   young     51    18        1.8
#> 31     clay     old     52     2        0.2
#> 32 concrete     old     52     4        0.4
#> 33    other     old     52     1        0.1
#> 34     clay   young     52     5        0.5
#> 35 concrete   young     52     3        0.3
#> 36    other   young     52     5        0.5
#> 37     clay     old     53     2        0.2
#> 38 concrete     old     53     1        0.1
fieldSummary(x, "age_cat")
#>   age_cat Count Percentage
#> 1     old   502       50.2
#> 2   young   498       49.8
fieldSummary(x, "material")
#>   material Count Percentage
#> 1     clay   354       35.4
#> 2 concrete   332       33.2
#> 3    other   314       31.4
fieldSummary(x, "material", lengthColumn = "length")
#>   material Count length Percentage
#> 1     clay   354  17537       35.4
#> 2 concrete   332  16457       33.2
#> 3    other   314  15533       31.4